+++ src/test/groovy/com/epam/reportportal/spock/description/UnrollThreeTestsDescriptionTest.java
@@ -1,0 +1,80 @@
+/*
+ * Copyright 2021 EPAM Systems
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.epam.reportportal.spock.description;
+
+import com.epam.reportportal.service.ReportPortal;
+import com.epam.reportportal.service.ReportPortalClient;
+import com.epam.reportportal.spock.ReportPortalSpockListener;
+import com.epam.reportportal.spock.features.HelloSpockSpecUnroll;
+import com.epam.reportportal.spock.utils.TestExtension;
+import com.epam.reportportal.spock.utils.TestUtils;
+import com.epam.reportportal.util.test.CommonUtils;
+import com.epam.ta.reportportal.ws.model.StartRQ;
+import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.runner.Result;
+import org.mockito.ArgumentCaptor;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static com.epam.reportportal.spock.utils.TestUtils.*;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.equalTo;
+import static org.mockito.ArgumentMatchers.same;
+import static org.mockito.Mockito.*;
+
+public class UnrollThreeTestsDescriptionTest {
+
+	private static final String STATIC_PART = "Expect: \n";
+
+	private final String classId = CommonUtils.namedId("class_");
+	private final List<String> methodIds = Stream.generate(() -> CommonUtils.namedId("method_")).limit(3).collect(Collectors.toList());
+
+	private final ReportPortalClient client = mock(ReportPortalClient.class);
+
+	@BeforeEach
+	public void setupMock() {
+		TestUtils.mockLaunch(client, null, classId, methodIds);
+		TestUtils.mockBatchLogging(client);
+		TestExtension.listener = new ReportPortalSpockListener(ReportPortal.create(client, standardParameters(), testExecutor()));
+	}
+
+	@Test
+	public void verify_static_test_code_reference_generation() {
+		Result result = runClasses(HelloSpockSpecUnroll.class);
+
+		assertThat(result.getFailureCount(), equalTo(0));
+
+		verify(client).startTestItem(any());
+		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
+		verify(client, times(3)).startTestItem(same(classId), captor.capture());
+
+		List<StartTestItemRQ> items = captor.getAllValues();
+		List<String> descriptions = items.stream().map(StartRQ::getDescription).collect(Collectors.toList());
+
+		assertThat(descriptions,
+				containsInAnyOrder(STATIC_PART + "name: Spock; length: 5",
+						STATIC_PART + "name: Kirk; length: 4",
+						STATIC_PART + "name: Scotty; length: 6"
+				)
+		);
+	}
+}